Converted Artist,Albums tabs in recycler view. now they use recyclerView. And Coordinator layout is implemented in respective fragments.
Artist,Albums tabs are converted to use coordinator layout is their
respective fragments and they now support the collapsing the appBar. The
playlist tabs still uses ListView and SoundCloud Tab is empty that is it
has a fragment with only a textView.
Gradle files have changed in this step. As android studio is updated to
newest version and newest gradle plugin was also needed.