Signed in user and got access token, signed in and then brought new fragment by removing previous login fragment, implemented shared prefs.
1)Included Java Api Wrapper ,Made a login Page, sent my default username
and password through a background and got the Acess Token.
2)Implemented shared preferences to save the loginstate and user
credentials and maintain the state that whether we need to login again
or not.
3)Implemented the logic to remove the soundCloud login fragment from the
viewpager at runtime and bring the new Fragment (that is empty at the
moment) which will be used to show the users content.
4)added searchView on the toolbar, have to implement its coded logic.
its yet to be done