Added bottom bar with a button in mainActivity, did some work on layout structure.
1)Added bottom bar to the main activity which shows current song playing
and its artist and has a button to play or pause this song.
2) made changes in layout, changed the color and design a little bit.
made sure that bottom bar with play pause button does not cover any of
the recyclerview items.
3) songplaybackAcitivity (The activity in which opens when we tap a song
item) , made changes in it, the mediaController ath the bottom covered 2
to 3 songs present in the recyclerview, made changes in layout so that
this does not happen. now songslist (RecylerView) stays above the
mediaController.
3) at this point when someone installs the application, first time it
requires the user to sign in to soundcloud. i have given my own id and
password by default. just pressing the sign in to soundcloud button
signs in to soundcloud, but application needs to be restarted than.
further on it works fine. one can search desired content of soundcloud.
this issue will be fixed in the upcoming commit.
4) also added a loader bar, while the query is running on soundcloud a
loader comes and says that query is running.